iommu/arm-smmu: don't enable SMMU device until probing has completed
BugLink: http://bugs.launchpad.net/bugs/1240072

commit fd90cecbde065eac6ecc3ef38abace725ad27010 upstream.

We currently reset and enable the SMMU before the device has finished
being probed, so if we fail later on (for example, because we couldn't
request a global irq successfully) then we will leave the device in an
active state.

This patch delays the reset and enabling of the SMMU hardware until
probing has completed.

Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>